requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.1
httpx==0.27.0
aiofiles==23.2.1
psutil==5.9.8
docker==7.0.0
playwright==1.43.0
selenium==4.20.0
pillow==10.3.0
redis==5.0.4
celery==5.4.0
prometheus-client==0.20.0
structlog==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.29
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.1
jinja2==3.1.3
markdown==3.6
pygments==2.18.0

# Additional AI/ML tools
openai==1.30.0
anthropic==0.28.0
google-generativeai==0.7.0
requests==2.32.0
beautifulsoup4==4.12.0
lxml==5.2.0
pandas==2.2.0
numpy==1.26.0
matplotlib==3.9.0

# Database & storage
sqlalchemy-utils==0.41.0
aioredis==2.0.1

# Monitoring & logging
sentry-sdk==2.0.0
